Lesson 8: Use *ngFor to list objects in component
First App - Lesson 8

In this lesson, you used the ngFor directive to repeat data dynamically in Angular templates. You also added a new array of data to be used in the Angular app. The application now dynamically renders a list of housing locations in the browser.

The app is taking shape, great job.